[Security Solution][Detection Engine] Remove RuleTypeSchema in favor of Type for TypeScript (#76586)
## Summary

Removes RuleTypeSchema in favor of Type for TypeScript. Does break out one function called `parseScheduleDates` into its own file to remove a circular ref issue.
