+++ x-pack/plugins/security_solution/public/detections/containers/detection_engine/rules/types.ts
@@ -6,7 +6,6 @@
 
 import * as t from 'io-ts';
 
-import { RuleTypeSchema } from '../../../../../common/detection_engine/types';
 import {
   author,
   building_block_type,
@@ -16,6 +15,7 @@ import {
   severity_mapping,
   timestamp_override,
   threshold,
+  type,
 } from '../../../../../common/detection_engine/schemas/common/schemas';
 import {
   listArray,
@@ -44,7 +44,7 @@ export const NewRuleSchema = t.intersection([
     name: t.string,
     risk_score: t.number,
     severity: t.string,
-    type: RuleTypeSchema,
+    type,
   }),
   t.partial({
     actions: t.array(action),
@@ -117,7 +117,7 @@ export const RuleSchema = t.intersection([
     severity: t.string,
     severity_mapping,
     tags: t.array(t.string),
-    type: RuleTypeSchema,
+    type,
     to: t.string,
     threat: t.array(t.unknown),
     updated_at: t.string,
